Revert "Rename the ray.operator module to ray.ray_operator (#13705)"
This reverts commit 9b7944c885bfa0cec230f2ef8a5cb7abb126533c.
