+++ python/setup.py
@@ -449,7 +449,7 @@ setuptools.setup(
             "ray=ray.scripts.scripts:main",
             "rllib=ray.rllib.scripts:cli [rllib]",
             "tune=ray.tune.scripts:cli",
-            "ray-operator=ray.ray_operator.operator:main",
+            "ray-operator=ray.operator.operator:main",
             "serve=ray.serve.scripts:cli",
         ]
     },
